Make the squirrel function self contained
