+++ main.js
@@ -6,6 +6,7 @@ const { createWindow, createMenu, getMainWindow, triggerStartupDialogs } = requi
 const { notify } = require('./js/notification');
 const { getUserPreferences } = require('./js/user-preferences.js');
 const i18n = require('./src/configs/i18next.config');
+const { handleSquirrelEvent } = require('./js/squirrel.js');
 
 /* eslint-disable */
 if (handleSquirrelEvent(app)) {
@@ -170,69 +171,3 @@ catch (_)
     // eslint-disable-next-line no-empty
     // We don't need to do anything in this block.
 }
-
-/* eslint-disable */
-function handleSquirrelEvent(application) {
-    if (process.argv.length === 1) {
-        return false;
-    }
-
-    const ChildProcess = require('child_process');
-    const path = require('path');
-
-    const appFolder = path.resolve(process.execPath, '..');
-    const rootAtomFolder = path.resolve(appFolder, '..');
-    const updateDotExe = path.resolve(path.join(rootAtomFolder, 'Update.exe'));
-    const exeName = path.basename(process.execPath);
-
-    const spawn = function(command, args) {
-        let spawnedProcess, error;
-
-        try {
-            spawnedProcess = ChildProcess.spawn(command, args, {
-                detached: true
-            });
-        } catch (error) {}
-
-        return spawnedProcess;
-    };
-
-    const spawnUpdate = function(args) {
-        return spawn(updateDotExe, args);
-    };
-
-    const squirrelEvent = process.argv[1];
-    switch (squirrelEvent) {
-        case '--squirrel-install':
-        case '--squirrel-updated':
-            // Optionally do things such as:
-            // - Add your .exe to the PATH
-            // - Write to the registry for things like file associations and
-            //   explorer context menus
-
-            // Install start menu shortcuts
-            spawnUpdate(['--createShortcut', exeName]);
-
-            setTimeout(application.quit, 1000);
-            return true;
-
-        case '--squirrel-uninstall':
-            // Undo anything you did in the --squirrel-install and
-            // --squirrel-updated handlers
-
-            // Remove start menu shortcuts
-            spawnUpdate(['--removeShortcut', exeName]);
-
-            setTimeout(application.quit, 1000);
-            return true;
-
-        case '--squirrel-obsolete':
-            // This is called on the outgoing version of your app before
-            // we update to the new version - it's the opposite of
-            // --squirrel-updated
-
-            application.quit();
-            return true;
-    }
-};
-/* eslint-enable */
